requirements.txt
streamlit
tensorflow
librosa
matplotlib
Pillow
numpy
